rsvpte: bind native forwarding ownership to tunnel generations
Replace tunnel forwarding without allowing obsolete generation teardown
to remove its successor. Withdraw native state on head-end shutdown and
retain independent reservation descriptor handling.

Carries inherited RSVP framing, state, admission, configured-route,
native MPLS owner and private-context changes required by these source
interfaces, with their canonical test migrations. Their presence is not
a complete RFC or interoperability claim.

Bounded native attempt 014 passes: generation replacement through a new
bypass, old-context removal, stale-path rejection, ordinary-route
controls and head-end shutdown cleanup. Durable proof and independent
source-closure review are under
plan/verification-evidence/2026-09-22-baseline/rsvp-kernel-preparation/.

Known red remains in canonical PPP duplicate declarations, unrelated
compilation/style lint findings and previously recorded test
scaffolding. Full RSVP timing, policy, affinity and absent-feature gaps
remain explicit. Kernel MTU options repair and IGP-publisher
documentation are not included.

+++ internal/plugins/rsvpte/events.go
@@ -17,7 +17,7 @@ type lSPEvent struct {
 	TunnelEndpoint string  `json:"tunnel-endpoint"`
 	TunnelID       uint16  `json:"tunnel-id"`
 	LSPID          uint16  `json:"lsp-id"`
-	Bandwidth      float32 `json:"bandwidth"`
+	Bandwidth      float64 `json:"bandwidth"`
 	State          string  `json:"state"`
 }
